+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Change Log
 
 ## [Unreleased]
-[Unreleased]: https://github.com/bitdriftlabs/capture-sdk/compare/v0.23.5...HEAD
+[Unreleased]: https://github.com/bitdriftlabs/capture-sdk/compare/v0.23.8...HEAD
 
 ### Both
 
@@ -15,7 +15,8 @@
 
 **Fixed**
 
-- Nothing yet!
+- Switching from the fixed session strategy to the activity session strategy now correctly resets
+  stored state. In practice this is only an issue during onboarding / local development.
 
 ### Android
 
@@ -44,6 +45,42 @@
 **Fixed**
 
 - Nothing yet!
+
+## [0.23.8]
+[0.23.8]: https://github.com/bitdriftlabs/capture-sdk/releases/tag/v0.23.8
+
+### iOS
+
+**Changed**
+
+- Reduced the chances of SDK work delaying app termination when the app is closing
+
+## [0.23.7]
+[0.23.7]: https://github.com/bitdriftlabs/capture-sdk/releases/tag/0.23.7
+
+### Both
+
+**Added**
+
+- Added the device's region format (`region_format`) to app metrics to crash reports in [ISO_3166-1 alpha-2](https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2)
+- Added public `clearEntityId` / `clearEntityID` APIs to unset the current entity identifier
+
+**Fixed**
+
+- Improved session strategy callback execution to avoid threading issues.
+
+### iOS
+
+**Added**
+
+- Added whether low power mode (`low_power_mode_enabled`) was enabled to app metrics in crash reports
+- Exposed `isTracingActive`
+
+### Android
+
+**Added**
+
+- Added support for Google Play SDK verification in Android releases.
 
 ## [0.23.6]
 [0.23.6]: https://github.com/bitdriftlabs/capture-sdk/releases/tag/0.23.6